fixed artifacts fetching to check checksum of existing file before downloading

+++ cct/module.py
@@ -281,11 +281,18 @@ class CctResource(object):
         logger.debug("Fetching %s as a resource for module %s" % (self.url, self.name))
 
         self.path = os.path.join(directory, self.filename)
+
+        if os.path.exists(self.path) and self.check_sum():
+            logger.debug("using cached artifact for %s" % self.name)
+            return
+
         try:
             urlrequest.urlretrieve(self.url, self.path)
         except Exception as ex:
             raise CCTError("Cannot download artifact from url %s, error: %s" % (self.url, ex))
-        self.check_sum()
+
+        if not self.check_sum():
+            raise CCTError("Resource from %s doenst match required chksum %s" % (self.url, self.chksum))
 
     def check_sum(self):
         hash = getattr(hashlib, self.chksum[:self.chksum.index(':')])()
@@ -294,7 +301,7 @@ class CctResource(object):
                 hash.update(block)
         if self.chksum[self.chksum.index(':') + 1:] == hash.hexdigest():
             return True
-        raise CCTError("Resource from %s doenst match required chksum %s" % (self.url, self.chksum))
+        return False
 
     def replace_variables(self, string):
         var_regex = re.compile('\$\{.*\}')
